requirements.txt
PyCryptodome
boto3
pytest